+++ examples/General/muteFromSearchUser.php
@@ -45,7 +45,7 @@ try {
     sleep($timeToSearch / 1000);
     $searchResponse = $ig->discover->search($usernameToFollow);
     $ig->event->sendNavigation('button', 'explore_popular', 'search_typeahead');
-    
+
     $searchResults = $searchResponse->getList();
     $rankToken = $searchResponse->getRankToken();
     $resultList = [];