+++ examples/Like/likeRecentMediaFromLocationSearch.php
@@ -51,7 +51,7 @@ try {
 
     // Get suggested searches and recommendations from Instagram.
     $ig->event->sendNavigation('button', 'explore_popular', 'search_typeahead');
-    
+
     $ig->event->sendNavigation('button', 'search_typeahead', 'serp_places');
 
     $locationItems = $ig->location->findPlaces($queryLocation)->getItems();